+++ examples/v1-next/apollo-federation-compatibility/CHANGELOG.md
@@ -1,5 +1,9 @@
 # example-apollo-federation-compatibility
 
+## 1.8.33
+
+## 1.8.32
+
 ## 1.8.31
 
 ## 1.8.30